requirements.txt
anyio==4.8.0
beautifulsoup4==4.13.3
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.9
dulwich==0.22.7
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
more-itertools==10.6.0
msgpack==1.1.0
packaging==24.2
pbs-installer==2025.2.12
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.1.1
poetry-core==2.1.1
pycparser==2.22
pyproject_hooks==1.2.0
pytest==8.3.5
python-dotenv==1.0.1
RapidFuzz==3.12.1
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
tomlkit==0.13.2
trove-classifiers==2025.1.15.22
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.29.2
xattr==1.1.4
zstandard==0.23.0
